Support discover/resolve on specific interfaces
Return the netHandle of the interface on which a service was discovered,
and allow resolving services on specific interfaces based on that
information.

This provides basic support to avoid the current situation where
services are discovered on an unknown interface, and resolved on a
possibly different, unspecified interface. mdnsresponder usage is
being deprecated, but the functionality can then be kept when
implementing the NsdManager API with its replacement.

Bug: 190249673
Test: atest NsdManagerTest
Change-Id: I382d460339b0e320aa1441164041f575813e6559