[ELR#1] Add excludeLocalRoutes in VirtualNetwork
This is a preparation change for separating routing for the
bypassable VPN network. The routing will need to be updated
depending on whether the local traffic is excluded in the VPN
network. This member is expected to be used in VPN network
only, so add it in the VirtualNetwork, instead of Network.

Bug: 184750836
Test: cd system/netd ; atest
Change-Id: I0aac077c0ab82df1da0cf7db82f99210485b6ff1