[ELR#4] Install/Remove local routes rules with the interfaces
Install/Remove the local routes rules with the life time of the
interfaces. The default networks will break into two tables:
the original one and the one for placing the local rules.
The local routing table is still unfunctional since it's empty
and no routes matched.

Sample routings(VPN connected with default network on WiFi):

20000:	from all fwmark 0x0/0x10000 lookup local_network
23000:	from all fwmark 0x64/0x1ffff iif lo lookup wlan0
24000:	from all fwmark 0x0/0x30000 iif lo uidrange 0-99999 lookup ipsec8
24000:	from all fwmark 0xc0066/0xcffff lookup ipsec8
25000:	from all fwmark 0x0/0x10000 lookup iif lo wlan0_local
27000:	from all fwmark 0x66/0xffff lookup wlan0

Bug: 184750836
Test: cd system/netd ; atest
Change-Id: Ia67ec04e8c133e832dd39c10977f542d9c5b9cda