move definition of xt_bpf program pin locations to system/netd
As this ends up built in to the netd binary, and should thus
not be pulling in mainline (and thus mutable) header files.

We could technically export the path out of the netd_updatable.so,
but there is no benefit to doing so.

This is because the pin directory is determined by selinux access
having to make these available to netd & iptables (netdutils_wrappers),
and the entire structure of the iptables setup (ie. chains, etc) is
already entirely left up to netd.

Thus the only meaningful thing mainline could do here is rename
the programs (while maintaining their behaviour) but that provides
absolutely no value.

(If the programs are missing, then bandwidth accounting fails,
and there is no way to add additional programs)

Bug: 235778483
Test: TreeHugger
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I3c235d63b3695deb04564bb929913a9050b9fc10